Optimize for non-matches
Testing textContent is very, very fast since it doesn't include all the overhead, but doesn't allow replacement. Using it to test if resources should be spent on the xpath query and multiple replace calls greatly improves performance if no matches are found, at a cost of a performance loss if there are matches. This is preferred firstly because a user typically expects there to be no performance impact if the script doesn't actually change anything, and secondly because the gains are very large compared to the losses in the other case.